uncomment Raphael.getSubpath() and change its parameter types from int to double since testing shows that it seems to work

+++ src/com/hydro4ge/raphaelgwt/client/Raphael.java
@@ -378,13 +378,9 @@ public class Raphael extends Widget {
       return ((RaphaelJS.Path)el).getPointAtLength(length);
     }
 
-    /**
-     * doesn't seem to work
-     *
-    public String getSubpath(int from, int to) {
+    public String getSubpath(double from, double to) {
       return ((RaphaelJS.Path)el).getSubpath(from, to);
     }
-     */
   }
 
 }
